Rename module to CustomTool
