+++ docs/storybook/src/frontstage/ToolSettings.stories.tsx
@@ -10,7 +10,7 @@ import { AppUiDecorator } from "../Decorators";
 import { Page } from "../AppUiStory";
 import { createFrontstage, removeProperty } from "../Utils";
 import { ToolSettingsStory } from "./ToolSettings";
-import { CustomTool } from "../tools/ToolSettingsProperties";
+import { CustomTool } from "../tools/CustomTool";
 
 const meta = {
   title: "Frontstage/ToolSettings",
@@ -23,16 +23,6 @@ const meta = {
     },
     layout: "fullscreen",
   },
-  argTypes: {
-    frontstages: removeProperty(),
-    onFrontstageActivated: removeProperty(),
-  },
-} satisfies Meta<typeof ToolSettingsStory>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
   args: {
     frontstages: [
       createFrontstage({
@@ -49,6 +39,18 @@ export const Default: Story = {
         hideToolSettings: false,
       }),
     ],
+  },
+  argTypes: {
+    frontstages: removeProperty(),
+    onFrontstageActivated: removeProperty(),
+  },
+} satisfies Meta<typeof ToolSettingsStory>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+  args: {
     onFrontstageActivated: async () => {
       IModelApp.tools.register(CustomTool, UiFramework.localizationNamespace);
       IModelApp.tools.run(CustomTool.toolId);
